requirements.txt
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
pydantic>=2.5.0,<3.0.0
python-dotenv>=1.0.0
loguru>=0.7.0
requests>=2.31.0
msal>=1.28.0
mcp>=1.0.0
python-dateutil>=2.9.0
uvicorn>=0.30.0
starlette>=0.38.0
httpx>=0.27.0

# Dev dependencies
pytest>=8.0.0
pytest-cov>=5.0.0
mypy>=1.8.0
ruff>=0.3.0
